+activeCheckForDATDev +sandwichSensor(Unf/F)udger
I just wrote the functions to do the sandwich fudges and the DAT active
checks, but I haven't made any migrations at all. This is to encourage
the automatic timestamp/timestamp diagnostic things. My idea is to use
the active hash, sort the keys, and use that to print out the data.