diagHeaderGenerator instead of activeChecks for Diag files
It was dumb, and lazy to use activeChecks for the diag header
generation, so i made a function to do it. Next step is to implement
Jesse-style speedups to all the first

also another == -> eq problem

Out of Occ changed savings amount due to no longer using DATSP in active
checks.